travis: fix clang build
configure fails to correctly recognize clang's C++11 capability.
Therefore, it does not append the -std=c++11 flag and falls back
to use C++ TR1 which does not compile. We work around that by
manually appending the aforementioned flag in the release flags
section.

In an upcoming commit this should be fixed. Moreover, configure
shall fail if it cannot identify a compiler's C++11 capability
instead of falling back (better user experience).